[MOVE/#65] ENUM 패키지 생성 및 이동
