+++ src/main/java/com/togedy/togedy_server_v2/domain/schedule/entity/enums/ScheduleType.java
@@ -1,4 +1,4 @@
-package com.togedy.togedy_server_v2.domain.schedule.entity;
+package com.togedy.togedy_server_v2.domain.schedule.entity.enums;
 
 import lombok.Getter;
