--- src/main/java/com/togedy/togedy_server_v2/domain/study/entity/StudyRole.java
@@ -1,5 +1,0 @@
-package com.togedy.togedy_server_v2.domain.study.entity;
-
-public enum StudyRole {
-    LEADER, MEMBER
-}
