--- src/main/java/com/togedy/togedy_server_v2/domain/study/entity/StudyTier.java
@@ -1,4 +1,0 @@
-package com.togedy.togedy_server_v2.domain.study.entity;
-
-public enum StudyTier {
-}
