+++ src/main/java/com/togedy/togedy_server_v2/domain/study/enums/StudyRole.java
@@ -1,0 +1,5 @@
+package com.togedy.togedy_server_v2.domain.study.enums;
+
+public enum StudyRole {
+    LEADER, MEMBER
+}
