+++ src/main/java/com/togedy/togedy_server_v2/domain/study/enums/StudyTier.java
@@ -1,0 +1,4 @@
+package com.togedy.togedy_server_v2.domain.study.enums;
+
+public enum StudyTier {
+}
